pyproject.toml
[tool.poetry]
name = "pandas-rag-langgraph"
version = "0.1.0"
description = ""
authors = ["vbarda <vadym@langchain.dev>"]
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.11,<3.13"
gdown = "^5.2.0"
langchain-core = "^0.2.5"
langchain-community = "^0.2.4"
langchain-text-splitters = "^0.2.1"
langgraph = "^0.1.0"
langchain-groq = "^0.1.6"
langchain-cohere = "^0.1.9"
langchain-pinecone = "^0.1.1"
llama-index-core = "^0.10.55" 
llama-index-readers-smart-pdf-loader = "^0.1.4"
llmsherpa = "^0.1.4"
streamlit = "^1.36.0"
watchdog = "^4.0.1"


[tool.poetry.group.dev.dependencies]
langgraph-cli = "^0.1.35"
langgraph-sdk = "^0.1.18"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
